#10 Bugfix: Pipes do not leave the level
Pipes were sent outside of the level by the else-block of the hit.distance-check.
They were sent backwards in the wrong direction, so that they would sometimes leave the level and are sent in the wrong direction a couple of times.
This is fixed now!

I also added the coding support for VSCode.